# jara-market-backend/jaraman/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Jaramarket API",
        "description": "Jaramarket Meals Marketplace API for Customer and Vendor mobile applications",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "L5_SWAGGER_CONST_HOST",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/payments/initiate": {
            "post": {
                "tags": [
                    "Customer"
                ],
                "summary": "Initiate Payment",
                "description": "Initialize a Paystack transaction for an order.",
                "operationId": "bada6b1b735567ed25c83a55afb9da96",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "order_id",
                                    "amount"
                                ],
                                "properties": {
                                    "order_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "amount": {
                                        "type": "number",
                                        "example": 1500
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment initialized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "authorization_url": {
                                                    "type": "string",
                                                    "example": "https://checkout.paystack.com/..."
                                                },
                                                "reference": {
                                                    "type": "string",
                                                    "example": "T123456789"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Initialization failed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/payments/verify": {
            "post": {
                "tags": [
                    "Customer"
                ],
                "summary": "Verify Payment",
                "description": "Verify a Paystack transaction and update order status to 'paid'.",
                "operationId": "37ba882160b977eb4b74f3a5262159ae",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "reference"
                                ],
                                "properties": {
                                    "reference": {
                                        "type": "string",
                                        "example": "T123456789"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment verified"
                    },
                    "400": {
                        "description": "Verification failed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/payments/history": {
            "get": {
                "tags": [
                    "Customer"
                ],
                "summary": "Payment History",
                "description": "Retrieve history of successful payments for the user.",
                "operationId": "e30b29390dd67afcec9dcea782c3f480",
                "responses": {
                    "200": {
                        "description": "History retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/auth/reset-password": {
            "post": {
                "tags": [
                    "Customer Authentication",
                    "Vendor Authentication"
                ],
                "summary": "Reset Password",
                "description": "Resets the user's password using the OTP and email.",
                "operationId": "76b42f2a6cb70bd1ddf05213fc81c605",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "otp",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "otp": {
                                        "type": "string",
                                        "example": "123456"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "NewPassword123!"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "NewPassword123!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password reset successful"
                    },
                    "400": {
                        "description": "Invalid OTP or request"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}